ISSUE-47: refactor to ZipArchive and bump dependencies (#48)
* Remove chumper/zipper, add min PHP 7.1 and bump Library versions

This also syncs now with frictionlessdata/tableschema: v0.2.0

* Refactor to use Vanilla \ZipArchive()

Adds ext-zip to composer.json and checks at Package Class level if the ZipArchive class is present if not Throw \Exception

* Fix Utils::removeDir

It was assuming always that a direct path was a dir. May not be the case and removes warning during tests

* Update .install_cs_fixer.sh

Updates php-cs-fixer to 2.18.2. This may be a problem for PHP 7.2 > since fixer is 7.1.* max

* Type Hinting Function arguments

* PHPDOCS and type hinting

Not every function has PHPDocs but we will get there

* $descriptor can be JSON or an PHP Object so skipping type hint for now

Add more info to PHPDOCS

* Data source can be also an array. The @param entry is wrong here

* Update BaseResource.php

Small PHPDOCS updates

* Coding Standards

Mostly Indentation and PHPDOCS and type hints

* Indent PHPDOCS

* Indentation

Indentation

* Fix Mock function signature

* add dev dependency to php_codesniffer and make code psr-2 compliant.

This passes php vendor/bin/phpcs --standard=psr2

* CI and composer scripts for phpcs/phpcbf

Just a test. Can roll back if this goes against the project's rules

* Move to Xenial. Trusty 7.4 has not zip

* Removing left over Debug statement

* kept ::isZipPresent as public static on Package.php, but called only when needed.

Also quite a few PHPDOCs indentations and `@throws`